Oh so many changes
Every service works with apache except the key_server.  The key_server
doesn't work simply because I havent tried it yet.  Sort of fixed an
issue with the euca stuff: lists all images but we need to figure out
a better way to do it.  More things are in the conf file.  Logging is
a bit more centralized but the glance/nova wsgi needs cleaning.
Can create keys to multiple clouds at once.